+++ ui/server.py
@@ -2,64 +2,21 @@
 Notes:
     async endpoints always run on the main thread. Without they run on the thread pool.
 """
-import json
-import traceback
-
-import sys
 import os
-import socket
-import picklescan.scanner
-import rich
-
-SD_DIR = os.getcwd()
-print('started in ', SD_DIR)
-
-SD_UI_DIR = os.getenv('SD_UI_PATH', None)
-sys.path.append(os.path.dirname(SD_UI_DIR))
-
-CONFIG_DIR = os.path.abspath(os.path.join(SD_UI_DIR, '..', 'scripts'))
-MODELS_DIR = os.path.abspath(os.path.join(SD_DIR, '..', 'models'))
-
-USER_UI_PLUGINS_DIR = os.path.abspath(os.path.join(SD_DIR, '..', 'plugins', 'ui'))
-CORE_UI_PLUGINS_DIR = os.path.abspath(os.path.join(SD_UI_DIR, 'plugins', 'ui'))
-UI_PLUGINS_SOURCES = ((CORE_UI_PLUGINS_DIR, 'core'), (USER_UI_PLUGINS_DIR, 'user'))
-
-STABLE_DIFFUSION_MODEL_EXTENSIONS = ['.ckpt', '.safetensors']
-VAE_MODEL_EXTENSIONS = ['.vae.pt', '.ckpt']
-HYPERNETWORK_MODEL_EXTENSIONS = ['.pt']
-
-OUTPUT_DIRNAME = "Stable Diffusion UI" # in the user's home folder
-TASK_TTL = 15 * 60 # Discard last session's task timeout
-APP_CONFIG_DEFAULTS = {
-    # auto: selects the cuda device with the most free memory, cuda: use the currently active cuda device.
-    'render_devices': 'auto', # valid entries: 'auto', 'cpu' or 'cuda:N' (where N is a GPU index)
-    'update_branch': 'main',
-    'ui': {
-        'open_browser_on_start': True,
-    },
-}
-APP_CONFIG_DEFAULT_MODELS = [
-    # needed to support the legacy installations
-    'custom-model', # Check if user has a custom model, use it first.
-    'sd-v1-4', # Default fallback.
-]
+import traceback
+import logging
+from typing import List, Union
 
 from fastapi import FastAPI, HTTPException
 from fastapi.staticfiles import StaticFiles
 from starlette.responses import FileResponse, JSONResponse, StreamingResponse
 from pydantic import BaseModel
-import logging
-#import queue, threading, time
-from typing import Any, Generator, Hashable, List, Optional, Union
 
-from sd_internal import Request, Response, task_manager
+from sd_internal import app, model_manager, task_manager
 
-app = FastAPI()
+print('started in ', app.SD_DIR)
 
-modifiers_cache = None
-outpath = os.path.join(os.path.expanduser("~"), OUTPUT_DIRNAME)
-
-os.makedirs(USER_UI_PLUGINS_DIR, exist_ok=True)
+server_api = FastAPI()
 
 # don't show access log entries for URLs that start with the given prefix
 ACCESS_LOG_SUPPRESS_PATH_PREFIXES = ['/ping', '/image', '/modifier-thumbnails']
@@ -74,132 +31,6 @@ class NoCacheStaticFiles(StaticFiles):
 
         return super().is_not_modified(response_headers, request_headers)
 
-app.mount('/media', NoCacheStaticFiles(directory=os.path.join(SD_UI_DIR, 'media')), name="media")
-
-for plugins_dir, dir_prefix in UI_PLUGINS_SOURCES:
-    app.mount(f'/plugins/{dir_prefix}', NoCacheStaticFiles(directory=plugins_dir), name=f"plugins-{dir_prefix}")
-
-def getConfig(default_val=APP_CONFIG_DEFAULTS):
-    try:
-        config_json_path = os.path.join(CONFIG_DIR, 'config.json')
-        if not os.path.exists(config_json_path):
-            return default_val
-        with open(config_json_path, 'r', encoding='utf-8') as f:
-            config = json.load(f)
-            if 'net' not in config:
-                config['net'] = {}
-            if os.getenv('SD_UI_BIND_PORT') is not None:
-                config['net']['listen_port'] = int(os.getenv('SD_UI_BIND_PORT'))
-            if os.getenv('SD_UI_BIND_IP') is not None:
-                config['net']['listen_to_network'] = ( os.getenv('SD_UI_BIND_IP') == '0.0.0.0' )
-            return config
-    except Exception as e:
-        print(str(e))
-        print(traceback.format_exc())
-        return default_val
-
-def setConfig(config):
-    print( json.dumps(config) )
-    try: # config.json
-        config_json_path = os.path.join(CONFIG_DIR, 'config.json')
-        with open(config_json_path, 'w', encoding='utf-8') as f:
-            json.dump(config, f)
-    except:
-        print(traceback.format_exc())
-
-    try: # config.bat
-        config_bat_path = os.path.join(CONFIG_DIR, 'config.bat')
-        config_bat = []
-
-        if 'update_branch' in config:
-            config_bat.append(f"@set update_branch={config['update_branch']}")
-
-        config_bat.append(f"@set SD_UI_BIND_PORT={config['net']['listen_port']}")
-        bind_ip = '0.0.0.0' if config['net']['listen_to_network'] else '127.0.0.1'
-        config_bat.append(f"@set SD_UI_BIND_IP={bind_ip}")
-
-        config_bat.append(f"@set test_sd2={'Y' if config.get('test_sd2', False) else 'N'}")
-
-        if len(config_bat) > 0:
-            with open(config_bat_path, 'w', encoding='utf-8') as f:
-                f.write('\r\n'.join(config_bat))
-    except:
-        print(traceback.format_exc())
-
-    try: # config.sh
-        config_sh_path = os.path.join(CONFIG_DIR, 'config.sh')
-        config_sh = ['#!/bin/bash']
-
-        if 'update_branch' in config:
-            config_sh.append(f"export update_branch={config['update_branch']}")
-
-        config_sh.append(f"export SD_UI_BIND_PORT={config['net']['listen_port']}")
-        bind_ip = '0.0.0.0' if config['net']['listen_to_network'] else '127.0.0.1'
-        config_sh.append(f"export SD_UI_BIND_IP={bind_ip}")
-
-        config_sh.append(f"export test_sd2=\"{'Y' if config.get('test_sd2', False) else 'N'}\"")
-
-        if len(config_sh) > 1:
-            with open(config_sh_path, 'w', encoding='utf-8') as f:
-                f.write('\n'.join(config_sh))
-    except:
-        print(traceback.format_exc())
-
-def resolve_model_to_use(model_name:str, model_type:str, model_dir:str, model_extensions:list, default_models=[]):
-    config = getConfig()
-
-    model_dirs = [os.path.join(MODELS_DIR, model_dir), SD_DIR]
-    if not model_name: # When None try user configured model.
-        # config = getConfig()
-        if 'model' in config and model_type in config['model']:
-            model_name = config['model'][model_type]
-    if model_name:
-        is_sd2 = config.get('test_sd2', False)
-        if model_name.startswith('sd2_') and not is_sd2: # temp hack, until SD2 is unified with 1.4
-            print('ERROR: Cannot use SD 2.0 models with SD 1.0 code. Using the sd-v1-4 model instead!')
-            model_name = 'sd-v1-4'
-
-        # Check models directory
-        models_dir_path = os.path.join(MODELS_DIR, model_dir, model_name)
-        for model_extension in model_extensions:
-            if os.path.exists(models_dir_path + model_extension):
-                return models_dir_path
-            if os.path.exists(model_name + model_extension):
-                # Direct Path to file
-                model_name = os.path.abspath(model_name)
-                return model_name
-    # Default locations
-    if model_name in default_models:
-        default_model_path = os.path.join(SD_DIR, model_name)
-        for model_extension in model_extensions:
-            if os.path.exists(default_model_path + model_extension):
-                return default_model_path
-    # Can't find requested model, check the default paths.
-    for default_model in default_models:
-        for model_dir in model_dirs:
-            default_model_path = os.path.join(model_dir, default_model)
-            for model_extension in model_extensions:
-                if os.path.exists(default_model_path + model_extension):
-                    if model_name is not None:
-                        print(f'Could not find the configured custom model {model_name}{model_extension}. Using the default one: {default_model_path}{model_extension}')
-                    return default_model_path
-    raise Exception('No valid models found.')
-
-def resolve_ckpt_to_use(model_name:str=None):
-    return resolve_model_to_use(model_name, model_type='stable-diffusion', model_dir='stable-diffusion', model_extensions=STABLE_DIFFUSION_MODEL_EXTENSIONS, default_models=APP_CONFIG_DEFAULT_MODELS)
-
-def resolve_vae_to_use(model_name:str=None):
-    try:
-        return resolve_model_to_use(model_name, model_type='vae', model_dir='vae', model_extensions=VAE_MODEL_EXTENSIONS, default_models=[])
-    except:
-        return None
-
-def resolve_hypernetwork_to_use(model_name:str=None):
-    try:
-        return resolve_model_to_use(model_name, model_type='hypernetwork', model_dir='hypernetwork', model_extensions=HYPERNETWORK_MODEL_EXTENSIONS, default_models=[])
-    except:
-        return None
-
 class SetAppConfigRequest(BaseModel):
     update_branch: str = None
     render_devices: Union[List[str], List[int], str, int] = None
@@ -209,9 +40,25 @@ class SetAppConfigRequest(BaseModel):
     listen_port: int = None
     test_sd2: bool = None
 
-@app.post('/app_config')
+class LogSuppressFilter(logging.Filter):
+    def filter(self, record: logging.LogRecord) -> bool:
+        path = record.getMessage()
+        for prefix in ACCESS_LOG_SUPPRESS_PATH_PREFIXES:
+            if path.find(prefix) != -1:
+                return False
+        return True
+
+# don't log certain requests
+logging.getLogger('uvicorn.access').addFilter(LogSuppressFilter())
+
+server_api.mount('/media', NoCacheStaticFiles(directory=os.path.join(app.SD_UI_DIR, 'media')), name="media")
+
+for plugins_dir, dir_prefix in app.UI_PLUGINS_SOURCES:
+    app.mount(f'/plugins/{dir_prefix}', NoCacheStaticFiles(directory=plugins_dir), name=f"plugins-{dir_prefix}")
+
+@server_api.post('/app_config')
 async def setAppConfig(req : SetAppConfigRequest):
-    config = getConfig()
+    config = app.getConfig()
     if req.update_branch is not None:
         config['update_branch'] = req.update_branch
     if req.render_devices is not None:
@@ -231,121 +78,48 @@ async def setAppConfig(req : SetAppConfigRequest):
     if req.test_sd2 is not None:
         config['test_sd2'] = req.test_sd2
     try:
-        setConfig(config)
+        app.setConfig(config)
 
         if req.render_devices:
-            update_render_threads()
+            app.update_render_threads()
 
         return JSONResponse({'status': 'OK'}, headers=NOCACHE_HEADERS)
     except Exception as e:
         print(traceback.format_exc())
         raise HTTPException(status_code=500, detail=str(e))
 
-def is_malicious_model(file_path):
-    try:
-        scan_result = picklescan.scanner.scan_file_path(file_path)
-        if scan_result.issues_count > 0 or scan_result.infected_files > 0:
-            rich.print(":warning: [bold red]Scan %s: %d scanned, %d issue, %d infected.[/bold red]" % (file_path, scan_result.scanned_files, scan_result.issues_count, scan_result.infected_files))
-            return True
-        else:
-            rich.print("Scan %s: [green]%d scanned, %d issue, %d infected.[/green]" % (file_path, scan_result.scanned_files, scan_result.issues_count, scan_result.infected_files))
-            return False
-    except Exception as e:
-        print('error while scanning', file_path, 'error:', e)
-    return False
+def update_render_devices_in_config(config, render_devices):
+    if render_devices not in ('cpu', 'auto') and not render_devices.startswith('cuda:'):
+        raise HTTPException(status_code=400, detail=f'Invalid render device requested: {render_devices}')
 
-known_models = {}
-def getModels():
-    models = {
-        'active': {
-            'stable-diffusion': 'sd-v1-4',
-            'vae': '',
-            'hypernetwork': '',
-        },
-        'options': {
-            'stable-diffusion': ['sd-v1-4'],
-            'vae': [],
-            'hypernetwork': [],
-        },
-    }
+    if render_devices.startswith('cuda:'):
+        render_devices = render_devices.split(',')
 
-    def listModels(models_dirname, model_type, model_extensions):
-        models_dir = os.path.join(MODELS_DIR, models_dirname)
-        if not os.path.exists(models_dir):
-            os.makedirs(models_dir)
+    config['render_devices'] = render_devices
 
-        for file in os.listdir(models_dir):
-            for model_extension in model_extensions:
-                if not file.endswith(model_extension):
-                    continue
-
-                model_path = os.path.join(models_dir, file)
-                mtime = os.path.getmtime(model_path)
-                mod_time = known_models[model_path] if model_path in known_models else -1
-                if mod_time != mtime:
-                    if is_malicious_model(model_path):
-                        models['scan-error'] = file
-                        return
-                known_models[model_path] = mtime
-
-                model_name = file[:-len(model_extension)]
-                models['options'][model_type].append(model_name)
-
-        models['options'][model_type] = [*set(models['options'][model_type])] # remove duplicates
-        models['options'][model_type].sort()
-
-    # custom models
-    listModels(models_dirname='stable-diffusion', model_type='stable-diffusion', model_extensions=STABLE_DIFFUSION_MODEL_EXTENSIONS)
-    listModels(models_dirname='vae', model_type='vae', model_extensions=VAE_MODEL_EXTENSIONS)
-    listModels(models_dirname='hypernetwork', model_type='hypernetwork', model_extensions=HYPERNETWORK_MODEL_EXTENSIONS)
-    # legacy
-    custom_weight_path = os.path.join(SD_DIR, 'custom-model.ckpt')
-    if os.path.exists(custom_weight_path):
-        models['options']['stable-diffusion'].append('custom-model')
-
-    return models
-
-def getUIPlugins():
-    plugins = []
-
-    for plugins_dir, dir_prefix in UI_PLUGINS_SOURCES:
-        for file in os.listdir(plugins_dir):
-            if file.endswith('.plugin.js'):
-                plugins.append(f'/plugins/{dir_prefix}/{file}')
-
-    return plugins
-
-def getIPConfig():
-    ips = socket.gethostbyname_ex(socket.gethostname())
-    ips[2].append(ips[0])
-    return ips[2]
-
-@app.get('/get/{key:path}')
+@server_api.get('/get/{key:path}')
 def read_web_data(key:str=None):
     if not key: # /get without parameters, stable-diffusion easter egg.
         raise HTTPException(status_code=418, detail="StableDiffusion is drawing a teapot!") # HTTP418 I'm a teapot
     elif key == 'app_config':
-        config = getConfig(default_val=None)
-        if config is None:
-            config = APP_CONFIG_DEFAULTS
-        return JSONResponse(config, headers=NOCACHE_HEADERS)
+        return JSONResponse(app.getConfig(), headers=NOCACHE_HEADERS)
     elif key == 'system_info':
-        config = getConfig()
+        config = app.getConfig()
         system_info = {
             'devices': task_manager.get_devices(),
-            'hosts': getIPConfig(),
+            'hosts': app.getIPConfig(),
+            'default_output_dir': os.path.join(os.path.expanduser("~"), app.OUTPUT_DIRNAME),
         }
         system_info['devices']['config'] = config.get('render_devices', "auto")
         return JSONResponse(system_info, headers=NOCACHE_HEADERS)
     elif key == 'models':
-        return JSONResponse(getModels(), headers=NOCACHE_HEADERS)
-    elif key == 'modifiers': return FileResponse(os.path.join(SD_UI_DIR, 'modifiers.json'), headers=NOCACHE_HEADERS)
-    elif key == 'output_dir': return JSONResponse({ 'output_dir': outpath }, headers=NOCACHE_HEADERS)
-    elif key == 'ui_plugins': return JSONResponse(getUIPlugins(), headers=NOCACHE_HEADERS)
+        return JSONResponse(model_manager.getModels(), headers=NOCACHE_HEADERS)
+    elif key == 'modifiers': return FileResponse(os.path.join(app.SD_UI_DIR, 'modifiers.json'), headers=NOCACHE_HEADERS)
+    elif key == 'ui_plugins': return JSONResponse(app.getUIPlugins(), headers=NOCACHE_HEADERS)
     else:
         raise HTTPException(status_code=404, detail=f'Request for unknown {key}') # HTTP404 Not Found
 
-@app.get('/ping') # Get server and optionally session status.
+@server_api.get('/ping') # Get server and optionally session status.
 def ping(session_id:str=None):
     if task_manager.is_alive() <= 0: # Check that render threads are alive.
         if task_manager.current_state_error: raise HTTPException(status_code=500, detail=str(task_manager.current_state_error))
@@ -372,38 +146,14 @@ def ping(session_id:str=None):
     response['devices'] = task_manager.get_devices()
     return JSONResponse(response, headers=NOCACHE_HEADERS)
 
-def save_model_to_config(ckpt_model_name, vae_model_name, hypernetwork_model_name):
-    config = getConfig()
-    if 'model' not in config:
-        config['model'] = {}
-
-    config['model']['stable-diffusion'] = ckpt_model_name
-    config['model']['vae'] = vae_model_name
-    config['model']['hypernetwork'] = hypernetwork_model_name
-
-    if vae_model_name is None or vae_model_name == "":
-        del config['model']['vae']
-    if hypernetwork_model_name is None or hypernetwork_model_name == "":
-        del config['model']['hypernetwork']
-
-    setConfig(config)
-
-def update_render_devices_in_config(config, render_devices):
-    if render_devices not in ('cpu', 'auto') and not render_devices.startswith('cuda:'):
-        raise HTTPException(status_code=400, detail=f'Invalid render device requested: {render_devices}')
-
-    if render_devices.startswith('cuda:'):
-        render_devices = render_devices.split(',')
-
-    config['render_devices'] = render_devices
-
-@app.post('/render')
+@server_api.post('/render')
 def render(req : task_manager.ImageRequest):
     try:
-        save_model_to_config(req.use_stable_diffusion_model, req.use_vae_model, req.use_hypernetwork_model)
-        req.use_stable_diffusion_model = resolve_ckpt_to_use(req.use_stable_diffusion_model)
-        req.use_vae_model = resolve_vae_to_use(req.use_vae_model)
-        req.use_hypernetwork_model = resolve_hypernetwork_to_use(req.use_hypernetwork_model)
+        app.save_model_to_config(req.use_stable_diffusion_model, req.use_vae_model, req.use_hypernetwork_model)
+        req.use_stable_diffusion_model = model_manager.resolve_ckpt_to_use(req.use_stable_diffusion_model)
+        req.use_vae_model = model_manager.resolve_vae_to_use(req.use_vae_model)
+        req.use_hypernetwork_model = model_manager.resolve_hypernetwork_to_use(req.use_hypernetwork_model)
+
         new_task = task_manager.render(req)
         response = {
             'status': str(task_manager.current_state), 
@@ -419,7 +169,7 @@ def render(req : task_manager.ImageRequest):
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-@app.get('/image/stream/{session_id:str}/{task_id:int}')
+@server_api.get('/image/stream/{session_id:str}/{task_id:int}')
 def stream(session_id:str, task_id:int):
     #TODO Move to WebSockets ??
     task = task_manager.get_cached_task(session_id, update_ttl=True)
@@ -433,7 +183,7 @@ def stream(session_id:str, task_id:int):
     #print(f'Session {session_id} opened live render stream {id(task.buffer_queue)}')
     return StreamingResponse(task.read_buffer_generator(), media_type='application/json')
 
-@app.get('/image/stop')
+@server_api.get('/image/stop')
 def stop(session_id:str=None):
     if not session_id:
         if task_manager.current_state == task_manager.ServerStates.Online or task_manager.current_state == task_manager.ServerStates.Unavailable:
@@ -446,7 +196,7 @@ def stop(session_id:str=None):
     task.error = StopAsyncIteration('')
     return {'OK'}
 
-@app.get('/image/tmp/{session_id}/{img_id:int}')
+@server_api.get('/image/tmp/{session_id}/{img_id:int}')
 def get_image(session_id, img_id):
     task = task_manager.get_cached_task(session_id, update_ttl=True)
     if not task: raise HTTPException(status_code=410, detail=f'Session {session_id} has not submitted a task.') # HTTP410 Gone
@@ -458,49 +208,17 @@ def get_image(session_id, img_id):
     except KeyError as e:
         raise HTTPException(status_code=500, detail=str(e))
 
-@app.get('/')
+@server_api.get('/')
 def read_root():
-    return FileResponse(os.path.join(SD_UI_DIR, 'index.html'), headers=NOCACHE_HEADERS)
+    return FileResponse(os.path.join(app.SD_UI_DIR, 'index.html'), headers=NOCACHE_HEADERS)
 
-@app.on_event("shutdown")
+@server_api.on_event("shutdown")
 def shutdown_event(): # Signal render thread to close on shutdown
     task_manager.current_state_error = SystemExit('Application shutting down.')
 
-# don't log certain requests
-class LogSuppressFilter(logging.Filter):
-    def filter(self, record: logging.LogRecord) -> bool:
-        path = record.getMessage()
-        for prefix in ACCESS_LOG_SUPPRESS_PATH_PREFIXES:
-            if path.find(prefix) != -1:
-                return False
-        return True
-logging.getLogger('uvicorn.access').addFilter(LogSuppressFilter())
-
-# Check models and prepare cache for UI open
-getModels()
-
-# Start the task_manager
-task_manager.default_model_to_load = resolve_ckpt_to_use()
-task_manager.default_vae_to_load = resolve_vae_to_use()
-task_manager.default_hypernetwork_to_load = resolve_hypernetwork_to_use()
-
-def update_render_threads():
-    config = getConfig()
-    render_devices = config.get('render_devices', 'auto')
-    active_devices = task_manager.get_devices()['active'].keys()
-
-    print('requesting for render_devices', render_devices)
-    task_manager.update_render_threads(render_devices, active_devices)
-
-update_render_threads()
+# Init the app
+model_manager.init()
+app.init()
 
 # start the browser ui
-def open_browser():
-    config = getConfig()
-    ui = config.get('ui', {})
-    net = config.get('net', {'listen_port':9000})
-    port = net.get('listen_port', 9000)
-    if ui.get('open_browser_on_start', True):
-        import webbrowser; webbrowser.open(f"http://localhost:{port}")
-
-open_browser()
+app.open_browser()
